One press downloads, verifies and installs
The dashboard card's primary action is "Download and install" and now does exactly
that. startDownload carries an autoInstall intent through the pipeline: the card
shows the transfer, verifies the APK, and hands straight to the system installer.
The READY_TO_INSTALL card is only reached when the system prompt was dismissed, the
install failed, or the "install unknown apps" permission is still missing. Stopping
there on the happy path asked the user to confirm the same decision twice — the
platform's own prompt is the confirmation that counts, and it has not moved.

The long label needs about 150dp, so it falls back to "Download" below 380dp of
screen width or above a 1.15 font scale, and the action row is a FlowRow with
single-line ellipsised labels so an overrun wraps instead of clipping.

The status card is now one component shared by the dashboard and the App updates
screen, since both run the identical sequence; only the dismiss X and the "View"
action differ between hosts. That also moves the install-permission round trip into
the card, so it works the same wherever the flow was started from.
